requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-debug-toolbar==2.0
django-haystack==2.8.1
django-redis==4.10.0
dnspython==1.16.0
eventlet==0.25.0
fdfs-client-py==1.2.6
greenlet==0.4.15
hiredis==1.0.0
idna==2.8
itsdangerous==1.1.0
jieba==0.39
kombu==4.6.1
monotonic==1.5
mutagen==1.42.0
mysqlclient==1.4.2.post1
Pillow==6.0.0
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
Whoosh==2.7.4
